Use all 6 octets of a MAC address in a hash set to store known addresses
Compress the 6 octets of a MAC address into 4 octets using a FNV1a-32
hash function, and store them in an open-coded hashset that uses a
technique similar to cuckoo hashing.

A hashset is comprised of up to 32 compressed addresses, in a chain;
every link in this chain has 136 bytes due to alignment.

BLE and WiFi are counted in separated hashsets as well, to increase
precision.

Although this approach significantly increases the amount of memory
used by libpax, if the microcontroller runs out of DRAM, it'll just
stop counting until libpax_counter_reset() is called and the hashsets
are cleared.  A ballpark figure that disregards the operating system
overhead puts this limit at around 77000 unique MAC addresses for a
DRAM of 320KB, which is a pretty high limit.

It should be possible to reduce memory usage by reducing precision by
using a 16-bit hashing function instead (FNV1a-32 could still be used
if both low and high 16-bit numbers are XORed together for instance),
and removing the count field in struct paxhashmap.  For my use case,
this has not proven to be necessary yet, so I haven't tested these
modifications.

Another possible modification would be, if allocating a new hashset node
fails, we replace one element in the last link in the chain with the new
one.  This would not change the number of MAC addresses in the set, but
would work in a manner similar to a LRU cache (if the address is scanned
again, we'll try to add it again to the set anyway).